+++ programs/drift/src/instructions/keeper.rs
@@ -21,6 +21,7 @@ use crate::controller::spot_balance::update_spot_balances;
 use crate::controller::token::{receive, send_from_program_vault};
 use crate::error::ErrorCode;
 use crate::ids::admin_hot_wallet;
+use crate::ids::dflow_mainnet_aggregator_4;
 use crate::ids::{jupiter_mainnet_3, jupiter_mainnet_4, jupiter_mainnet_6, serum_program};
 use crate::instructions::constraints::*;
 use crate::instructions::optional_accounts::get_revenue_share_escrow_account;
@@ -1103,8 +1104,7 @@ pub fn handle_settle_pnl<'c: 'info, 'info>(
             &mut oracle_map,
             state,
             &clock,
-        )
-        .map(|_| ErrorCode::InvalidOracleForSettlePnl)?;
+        )?;
 
         controller::pnl::settle_pnl(
             market_index,
@@ -1118,8 +1118,7 @@ pub fn handle_settle_pnl<'c: 'info, 'info>(
             state,
             None,
             SettlePnlMode::MustSettle,
-        )
-        .map(|_| ErrorCode::InvalidOracleForSettlePnl)?;
+        )?;
     }
 
     if state.builder_codes_enabled() || state.builder_referral_enabled() {
@@ -1219,8 +1218,7 @@ pub fn handle_settle_multiple_pnls<'c: 'info, 'info>(
                 &mut oracle_map,
                 state,
                 &clock,
-            )
-            .map(|_| ErrorCode::InvalidOracleForSettlePnl)?;
+            )?;
 
             controller::pnl::settle_pnl(
                 *market_index,
@@ -1234,8 +1232,7 @@ pub fn handle_settle_multiple_pnls<'c: 'info, 'info>(
                 state,
                 Some(meets_margin_requirement),
                 mode,
-            )
-            .map(|_| ErrorCode::InvalidOracleForSettlePnl)?;
+            )?;
         }
 
         if state.builder_codes_enabled() || state.builder_referral_enabled() {
@@ -1721,6 +1718,7 @@ pub fn handle_liquidate_spot_with_swap_begin<'c: 'info, 'info>(
                     jupiter_mainnet_3::ID,
                     jupiter_mainnet_4::ID,
                     jupiter_mainnet_6::ID,
+                    dflow_mainnet_aggregator_4::ID,
                 ];
                 validate!(
                     whitelisted_programs.contains(&ix.program_id),