+++ sdk/src/isomorphic/grpc.node.ts
@@ -4,30 +4,34 @@ import type {
 	SubscribeUpdate,
 } from '@triton-one/yellowstone-grpc';
 import { CommitmentLevel } from '@triton-one/yellowstone-grpc';
-import { ClientDuplexStream, ChannelOptions } from '@grpc/grpc-js';
+import type { ClientDuplexStream, ChannelOptions } from '@grpc/grpc-js';
 
 import {
 	CommitmentLevel as LaserCommitmentLevel,
 	subscribe as LaserSubscribe,
+	CompressionAlgorithms,
+} from 'helius-laserstream';
+import type {
 	LaserstreamConfig,
 	SubscribeRequest as LaserSubscribeRequest,
 	SubscribeUpdate as LaserSubscribeUpdate,
-	CompressionAlgorithms,
 } from 'helius-laserstream';
 
 export {
-	ClientDuplexStream,
-	ChannelOptions,
-	SubscribeRequest,
-	SubscribeUpdate,
 	CommitmentLevel,
 	Client,
 	LaserSubscribe,
 	LaserCommitmentLevel,
+	CompressionAlgorithms,
+};
+export type {
+	ClientDuplexStream,
+	ChannelOptions,
+	SubscribeRequest,
+	SubscribeUpdate,
 	LaserstreamConfig,
 	LaserSubscribeRequest,
 	LaserSubscribeUpdate,
-	CompressionAlgorithms,
 };
 
 // Export a function to create a new Client instance